01_inverse_scaling/requirements.txt
# Core ML/AI
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
datasets>=2.14.0
tokenizers>=0.14.0

# API Clients
openai>=1.3.0
anthropic>=0.7.0
huggingface-hub>=0.19.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Dashboard
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Logging & Monitoring
wandb>=0.15.0
tensorboard>=2.14.0
tqdm>=4.66.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
requests>=2.31.0
aiohttp>=3.8.0
tenacity>=8.2.0
python-dateutil>=2.8.0
psutil>=5.9.0
